requirements.txt
absl-py==2.1.0
aiogram==2.22.1
aiohttp==3.8.1
aioice==0.9.0
aiortc==1.9.0
aiosignal==1.2.0
aiosmtplib==3.0.2
alabaster==0.7.13
albucore==0.0.17
albumentations==1.4.18
alembic==1.14.0
altair==5.4.1
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
av==12.3.0
Babel==2.10.0
backcall==0.2.0
backports-datetime-fromisoformat==2.0.3
bcrypt==4.3.0
beanie==1.29.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.9.0
cachetools==5.3.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
cmake==3.30.5
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==44.0.2
cycler==0.12.1
Cython==3.0.11
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
databases==0.9.0
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
Django==4.2.4
django-bootstrap5==23.3
djangorestframework==3.14.0
dnspython==2.7.0
docutils==0.19
easydict==1.13
ecdsa==0.19.1
email_validator==2.2.0
environs==14.1.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
exceptiongroup==1.2.2
execnet==2.1.1
executing==1.2.0
factor_analyzer==0.5.1
fastapi==0.115.6
fastapi-cli==0.0.7
fastapi-mail==1.4.2
fastapi_cors==0.0.6
fastjsonschema==2.16.3
filelock==3.16.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.3.1
fsspec==2024.10.0
gast==0.6.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.11.1
google-auth==2.23.0
google-cloud-speech==2.21.0
google-crc32c==1.6.0
google-pasta==0.2.0
googleapis-common-protos==1.60.0
greenlet==3.1.1
grpcio==1.58.0
grpcio-status==1.58.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
imageio==2.36.0
imageio-ffmpeg==0.5.1
imagesize==1.4.1
imutils==0.5.4
iniconfig==2.0.0
insightface==0.7.3
ipykernel==6.29.5
ipython==8.28.0
ipython-genutils==0.2.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.35
jaxlib==0.4.35
jedi==0.19.1
Jinja2==3.1.5
joblib==1.2.0
json5==0.9.25
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.3.0
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaleido==0.2.1
keras==3.6.0
Kivy==2.1.0
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.gstreamer==0.3.3
kivy-deps.sdl2==0.4.5
Kivy-Garden==0.1.5
kiwisolver==1.4.4
lazy-model==0.2.0
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.39.1
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.14
mistune==2.0.5
ml-dtypes==0.4.1
motor==3.7.0
moviepy==1.0.3
mpmath==1.2.1
multidict==6.0.2
namex==0.0.8
narwhals==1.10.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.1
onnx==1.17.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
pillow==11.1.0
platformdirs==3.2.0
pluggy==1.5.0
pocketsphinx==5.0.2
prettytable==3.11.0
proglog==0.1.10
prometheus-client==0.16.0
prompt_toolkit==3.0.48
proto-plus==1.22.3
protobuf==4.25.5
psutil==5.9.4
psycopg2==2.9.10
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.4.8
pyasn1-modules==0.3.0
PyAudio==0.2.13
pycparser==2.22
pydantic==2.10.5
pydantic-extra-types==2.10.1
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.0.0
pygame==2.1.2
Pygments==2.19.1
PyJWT==2.9.0
pylibsrtp==0.10.0
pymongo==4.12.0
PyMuPDF==1.24.13
pyOpenSSL==24.2.1
pyparsing==3.0.9
pypdf==5.3.0
pypiqe==1.2
pypiwin32==223
pyrsistent==0.19.3
PySnooper==1.2.1
pyTelegramBotAPI==4.7.1
pytest==8.3.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.10.1
python-jose==3.4.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-socketio==5.11.4
pytz==2023.3
pywin32==304
pywinpty==2.0.10
PyYAML==6.0.2
pyzmq==25.0.2
redis==5.1.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.20.0
rsa==4.9.1
schemdraw==0.16
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sounddevice==0.4.6
soupsieve==2.4
SpeechRecognition==3.10.0
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.37
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.41.3
streamlit==1.39.0
streamlit-webrtc==0.47.9
sympy==1.13.1
tenacity==8.2.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-cpu==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2024.9.20
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.2
torch==2.5.0
torchvision==0.20.0
tornado==6.2
tqdm==4.65.0
trace-updater==0.0.9.1
traitlets==5.14.3
typer==0.15.1
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.10.0
uri-template==1.2.0
urllib3==1.26.12
uvicorn==0.34.0
watchdog==5.0.3
watchfiles==1.0.4
wcwidth==0.2.6
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==2.3.6
widgetsnbextension==4.0.13
wrapt==1.15.0
wsproto==1.2.0
xxhash==3.2.0
yarl==1.8.1
zipp==3.15.0
